Refactor to support custom calculator-part1 (#278)
* Change policy submodule name to schedulers

* Rename daily_schedule.py to schedule.py
